+++ rust/src/api/mod.rs
@@ -1,1 +1,2 @@
+pub mod nexus;
 pub mod simple;